Refactor filenames/organization of external services
