+++ server/service/spotify.js
@@ -1,5 +1,5 @@
 const Promise = require('bluebird');
-const config = require('./config');
+const config = require('../config');
 
 const axios = require('axios');
 const querystring = require('querystring');
