+++ test/musixMatchSpec.js
@@ -1,17 +1,17 @@
-const musixMatchHelpers = require('../server/musixMatchHelpers.js');
+const musixMatch = require('../server/service/musixMatch.js');
 const expect = require('chai').expect;
 
 describe('getLyricsByTrackId', function() {
 
   it('function should exist', function() {
-    expect(musixMatchHelpers.getLyricsByTrackId).to.be.a('function');
+    expect(musixMatch.getLyricsByTrackId).to.be.a('function');
   });
 
   const testTrackId = '15445219';
   let queryData = {};
 
   it('should take string', function() {
-    return musixMatchHelpers.getLyricsByTrackId(testTrackId)
+    return musixMatch.getLyricsByTrackId(testTrackId)
     .then(data => {
       queryData = data;
     });
@@ -30,7 +30,7 @@ describe('getLyricsByTrackId', function() {
 describe('getLyricsByTitleAndArtist', function() {
 
   it('function should exist', function() {
-    expect(musixMatchHelpers.getLyricsByTitleAndArtist).to.be.a('function');
+    expect(musixMatch.getLyricsByTitleAndArtist).to.be.a('function');
   });
 
   const testTitle = 'like a rolling stone';
@@ -38,7 +38,7 @@ describe('getLyricsByTitleAndArtist', function() {
   let queryData = {};
 
   it('should take 2 string inputs, title and artist', function() {
-    return musixMatchHelpers.getLyricsByTitleAndArtist(testTitle, testArtist)
+    return musixMatch.getLyricsByTitleAndArtist(testTitle, testArtist)
     .then(data => {
       queryData = data;
     });
@@ -57,28 +57,28 @@ describe('getLyricsByTitleAndArtist', function() {
 describe('searchByTitleAndArtist', function() {
 
   it('function should exist', function() {
-    expect(musixMatchHelpers.searchByTitleAndArtist).to.be.a('function');
+    expect(musixMatch.searchByTitleAndArtist).to.be.a('function');
   });
-  
+
   const testTitle = 'i cant get no satisfaction';
   const testArtist = 'the rolling stones';
 
   it('should take only title and return array in track_list property', function() {
-    return musixMatchHelpers.searchByTitleAndArtist(testTitle, '')
+    return musixMatch.searchByTitleAndArtist(testTitle, '')
     .then(data => {
       expect(data.track_list).to.be.a('array');
     });
   });
 
   it('should take only artist and return array in track_list property', function() {
-    return musixMatchHelpers.searchByTitleAndArtist('', testArtist)
+    return musixMatch.searchByTitleAndArtist('', testArtist)
     .then(data => {
       expect(data.track_list).to.be.a('array');
     });
   });
 
   it('should take both title and return array in track_list property', function() {
-    return musixMatchHelpers.searchByTitleAndArtist(testTitle, testArtist)
+    return musixMatch.searchByTitleAndArtist(testTitle, testArtist)
     .then(data => {
       expect(data.track_list).to.be.a('array');
     });
