+++ test/spotifySpec.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
 
 const rewire = require('rewire');
-const spotifyHelpers = rewire('../server/spotifyHelpers');
+const spotify = rewire('../server/service/spotify');
 
 const expect = chai.expect;
 chai.use(chaiAsPromised);
@@ -22,7 +22,7 @@ describe( 'spotify API:', function() {
       const testTitle = 'like a rolling stone';
       const testArtist = 'the rolling stones';
 
-      return expect(spotifyHelpers.getSongByTitleAndArtist(testTitle, testArtist)).to.eventually.equal('spotify:track:0oXnmfo2kW3joSeiXoazdV');
+      return expect(spotify.getSongByTitleAndArtist(testTitle, testArtist)).to.eventually.equal('spotify:track:0oXnmfo2kW3joSeiXoazdV');
     });
 
   });
