+++ test/stubs.js
